Refactor: Restructured Project Architecture
