+++ .gitignore
@@ -82,6 +82,9 @@ target/
 profile_default/
 ipython_config.py
 
+# Profiling
+profiling.prof
+
 # pyenv
 #   For a library or package, you might want to ignore these files since the code is
 #   intended to run in multiple environments; otherwise, check them in:
